Optimize day23 puzzle
Instead of keeping the cups in the order provided to us, reorder them
to be a sorted, sequential array.

If we do that, then we don't need the lookup table anymore: a cups
address becomes directly correlated to its value. This in turn cuts
out 10 million hashes, which speeds up the solution substantially.