Improved day11 solution
Rather than considering every single galaxy one at a time, we can
calculate the x deltas and the y deltas independently.

When we do that, since multiple galaxies are always on the same row or
column, we can batch the calculations together and vastly reduce the
number of calculations needed.