Optimize day01 solution
Fold the right list into pairs; value and frequency.

Then rather than iterating through the left list, iterate through the
right list, keeping a cursor on the left list. We advance the cursor
whenever the left value is below the right value. If it's equal, we
add to the accumulator. If it's over, we try the next value.

This cuts down the number of values we have to test.